Support setting the route without navigating
Necessary for mutating route specific data without causing navigation changes.
The desired usage here would be that `oldState.map { $0.instanceIdentifier } == newState.map { $0.instanceIdentifier}`, and has significant side effects (described below) if not the case.

Note this carries a risk of having unintended effects if the user uses this incorrectly as it could cause the UI & navigationState to go out-of-sync if one changes the actual segments in the route.

However, the above side effect also Fixes #26, in that one can use this to react to UIKit based changes to the route. (IE system UINavigationBar back button)