Provide more information on Routable fatalErrors
The default implementation for the function to push, pon, and change
provided via protocol extension fatalErrors to notify developers that
they haven't provided the required implementation.

The message used in the fatalError was not informative enough. These
changes make it granular to the actual function being used, as well as
adding the type that caused the failure, and the route element
parameters that the code was trying to act on.

The new error message for push for example will look something like:

```
This routable cannot push segments. You have not implemented it.
(Asked RootRoutable to push landing)
```